refactor(gst_exporter,color_grading,file_manager,execution_order,histogram): update GStreamer API calls, fix type conversions, and remove dead code
- Changed bus_watch_id type from Option<SourceId> to Option<gst::BusWatchGuard> in GstExporter
- Removed pipeline.set_timeline() call in GstExporter::export()
- Changed query_position() from Result to Option return type handling
- Changed add_profile() calls to use references instead of upcast_ref()
- Added Ok(()) returns to configure_lut() and setup

+++ src-tauri/crates/aether_core/src/scopes/histogram/analysis.rs
@@ -56,18 +56,18 @@ impl HistogramAnalyzer {
         analysis.midtone_percentage = (midtone_samples as f32 / total_samples as f32) * 100.0;
         analysis.highlight_percentage = (highlight_samples as f32 / total_samples as f32) * 100.0;
 
-        analysis.black_clipped = luma_data[0] > total_samples / 1000;
-        analysis.white_clipped = luma_data[255] > total_samples / 1000;
+        analysis.black_clipped = (luma_data[0] as u64) > total_samples / 1000;
+        analysis.white_clipped = (luma_data[255] as u64) > total_samples / 1000;
 
-        let mut min_bin = 255;
-        let mut max_bin = 0;
+        let mut min_bin = 255usize;
+        let mut max_bin = 0usize;
         for (bin, &count) in luma_data.iter().enumerate() {
             if count > 0 {
                 min_bin = min_bin.min(bin);
                 max_bin = max_bin.max(bin);
             }
         }
-        analysis.dynamic_range = max_bin - min_bin;
+        analysis.dynamic_range = (max_bin - min_bin) as u8;
 
         Ok(analysis)
     }
